requirements.txt
# ===================================
# 美股自选股智能分析系统 - 依赖列表
# ===================================

# 核心依赖
python-dotenv>=1.0.0
tenacity>=8.2.0
sqlalchemy>=2.0.0
schedule>=1.2.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
yfinance>=0.2.0

# AI 与 LLM
litellm>=1.0.0
tiktoken>=0.5.0
json-repair>=0.25.0

# 其它工具
PyYAML>=6.0.0
jinja2>=3.1.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
markdown2>=2.4.0
fake-useragent>=1.4.0
httpx[socks]>=0.26.0
